Refactoring Package Structure
