+++ component/src/test/java/org/wso2/extension/siddhi/execution/timeseries/KernelExtensionTestCase.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org)
- * All Rights Reserved.
+ * Copyright (c)  2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -12,12 +11,12 @@
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  */
 
-package org.wso2.siddhi.extension.timeseries;
+package org.wso2.extension.siddhi.execution.timeseries;
 
 import junit.framework.Assert;
 import org.apache.log4j.Logger;
