+++ component/src/test/java/org/wso2/extension/siddhi/execution/timeseries/LinearRegressionTestcase.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c)  2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.siddhi.extension.timeseries;
+package org.wso2.extension.siddhi.execution.timeseries;
 
 import org.apache.log4j.Logger;
 import org.junit.Assert;
