+++ feature/pom.xml
@@ -14,27 +14,31 @@
  ~ See the License for the specific language governing permissions and
  ~ limitations under the License.
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
+
     <parent>
-        <groupId>org.wso2.extension.siddhi.window</groupId>
-        <artifactId>org.wso2.extension.siddhi.window</artifactId>
+        <groupId>org.wso2.extension.siddhi.execution</groupId>
+        <artifactId>siddhi-execution-timeseries</artifactId>
         <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../component/pom.xml</relativePath>
+        <relativePath>../pom.xml</relativePath>
     </parent>
-    <artifactId>org.wso2.extension.siddhi.window.feature</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
+
+    <artifactId>org.wso2.extension.siddhi.execution.timeseries.feature</artifactId>
     <packaging>pom</packaging>
-    <name>WSO2 Carbon - Extension Unique Time Feature</name>
+    <name>WSO2 Siddhi Extension - Time Series Feature</name>
     <url>http://wso2.org</url>
-    <description>This feature contains extension feature for unique time</description>
+    <description>This feature contains Siddhi extension feature for Time Series.</description>
+
     <dependencies>
         <dependency>
-            <groupId>org.wso2.extension.siddhi.window</groupId>
-            <artifactId>org.wso2.extension.siddhi.window.uniquetime</artifactId>
+            <groupId>org.wso2.extension.siddhi.execution</groupId>
+            <artifactId>org.wso2.extension.siddhi.execution.timeseries</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
+
     <build>
         <plugins>
             <plugin>
@@ -49,7 +53,7 @@
                             <goal>p2-feature-gen</goal>
                         </goals>
                         <configuration>
-                            <id>org.wso2.extension.siddhi.window.uniquetime</id>
+                            <id>org.wso2.extension.siddhi.execution.timeseries</id>
                             <propertiesFile>feature.properties</propertiesFile>
                             <adviceFile>
                                 <properties>
@@ -59,7 +63,7 @@
                             </adviceFile>
                             <bundles>
                                 <bundleDef>
-                                    org.wso2.extension.siddhi.window:org.wso2.extension.siddhi.window.uniquetime
+                                    org.wso2.extension.siddhi.execution:org.wso2.extension.siddhi.execution.timeseries
                                 </bundleDef>
                             </bundles>
                         </configuration>
