+++ release.properties
@@ -7,6 +7,6 @@ preparationGoals=clean install
 scm.id=my-scm-server
 remoteTagging=true
 projectVersionPolicyId=default
-scm.commentPrefix=[maven-release-plugin] 
+scm.commentPrefix=[maven-release-plugin]
 exec.snapshotReleasePluginAllowed=false
 completedPhase=check-poms
